project v0.0.1

+++ src/pages/Home.jsx
@@ -20,33 +20,36 @@ const Home = () => {
         
     }
     const addItem = async(e) => {
-        if(name!==""){
+        if(name!=="" && !name.includes(" ")){
             e.preventDefault();
             setItems([...items,name])
             console.log(items)
             setname("")
-            
-            var ref = await db.getRef().collection("Management")
 
+            var ref = await db.getRef().collection("Management")
             ref.doc(name).set({
                 ItemsList:[]
             })
 
         }else{
-            window.alert("empty input")
+            window.alert("empty input or space cannot be used as a input.")
         }
         return true;
     }
 
     const deleteItem = async (e,id) => {
         e.preventDefault()
-        var ref = await db.getRef().collection("Management")
-        ref.doc(items[id]).delete();
-        setItems(
-            items.filter( (v,i) => {
-                return (i!==id)
-            })
-        )
+        if(window.confirm("are you sure you want to delete : "+ items[id])){
+            var ref = await db.getRef().collection("Management")
+            ref.doc(items[id]).delete();
+            setItems(
+                items.filter( (v,i) => {
+                    return (i!==id)
+                })
+            )    
+        }else{
+            window.alert("delete operation cancelled")
+        }
         
        return true;
     }
@@ -72,13 +75,14 @@ const Home = () => {
             </div>
             
             <div className="app-list-items">
+                
                 {
                     items.map( (ele,i) => {
                         return(
-                            <div className="app-item" key={i+""}>
+                            <div className="app-item" key={(i)+""}>
                                 <div className="seminar-name">{ele}</div>
-                                <button className="btn btn-dark btn-block" onClick={(e) => deleteItem(e,i)}>DELETE</button>
-                                <button className="btn btn-dark btn-block" ><a href={"/Welcome/:"+ele}  target="_blank">OPEN</a></button>
+                                <button className="btn btn-dark btn-block" onClick={(e) => deleteItem(e,i)} >DELETE</button>
+                                <button className="btn btn-dark btn-block" style={{margin:"20px"}}><a style={{textDecoration:"none",color:"white",margin:"0px"}} href={"/Welcome/"+ele}  target="_blank" rel="noreferrer">OPEN</a></button>
                             </div>
                         )
                     })
